solution2/requirements.txt
opencv-python>=4.5.0
numpy>=1.19.0
matplotlib>=3.3.0
open3d>=0.13.0
pyyaml>=5.4.0